refactor(naming): rename Departure files to StopTime
Rename files that represent stop time events (both arrivals and
departures in GTFS) from "Departure" to "StopTime" to better reflect
that they are not limited to departure-only data.

This commit only moves files and updates import paths; internal
identifiers (function names, interface names, variables) are kept
as-is and will be renamed in follow-up commits.

- use-nearby-departures.ts -> use-nearby-stop-times.ts
- departure-views.ts -> stop-time-views.ts
- departure-item.tsx -> stop-time-item.tsx
- flat-departure-item.tsx -> flat-stop-time-item.tsx
- (matching .stories.tsx and .test.ts files)
