+++ src/app.tsx
@@ -8,7 +8,7 @@ import type { LoadResult } from './repositories/athenai-repository-v2';
 import { useTransitRepository } from './hooks/use-transit-repository';
 import { useUserSettings } from './hooks/use-user-settings';
 import { useDateTime } from './hooks/use-date-time';
-import { useNearbyDepartures } from './hooks/use-nearby-departures';
+import { useNearbyDepartures } from './hooks/use-nearby-stop-times';
 import { useSelection } from './hooks/use-selection';
 import { useStopHistory } from './hooks/use-stop-history';
 import { useAnchors } from './hooks/use-anchors';
