+++ src/components/bottom-sheet-header.tsx
@@ -4,7 +4,7 @@ import type { Agency } from '../types/app/transit';
 import type { DepartureViewMeta } from '../types/app/transit-composed';
 import type { NearbyStopsCounts } from './bottom-sheet';
 import { DEFAULT_AGENCY_LANG } from '../config/transit-defaults';
-import { DEPARTURE_VIEWS } from '../domain/transit/departure-views';
+import { DEPARTURE_VIEWS } from '../domain/transit/stop-time-views';
 import { getAgencyDisplayNames } from '../domain/transit/get-agency-display-name';
 import { createLogger } from '../lib/logger';
 import { routeTypeColor } from '../utils/route-type-color';
